% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewrite_gro.R
\name{rewrite_gro}
\alias{rewrite_gro}
\title{Rewrites the SHAW gro file based on snow melt timing}
\usage{
rewrite_gro(
  model_dir,
  frost = paste0(model_dir, "/out/frost.txt"),
  gro_file = paste0(model_dir, "/", basename(model_dir), ".gro"),
 
    model = "/Volumes/research_storage2/arctic/arctic_point_modeling/SHAW/hourly/calibration_data/phenology_gam.rds",
  on_duration = 30,
  write_file = F
)
}
\arguments{
\item{model_dir}{This is the directory where all inputs will be written.}

\item{frost}{File with snow outputs}

\item{gro_file}{Name of gro file}

\item{model}{Some kind of model object wit h SDD and lai_start}

\item{on_duration}{Period of time over which LAI ramps up from 0 to max, in days}

\item{write_file}{Only writes file (to same location) if set to true.}
}
\description{
Takes output frost.txt, cleans, and calculates SDD for each water year.
Given a model with sdd as input, calculates first day of leaf out.
}
\details{
This function allows you to build SHAW files based on desired parameters.
}
\examples{
observed <- data.frame(depth = c(0.1, 0.5, 1), VWC = c(0.4, 0.2, 0.1))
}
\keyword{SHAW.}
\keyword{hydrology,}
