% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shaw_inp.R
\name{shaw_inp}
\alias{shaw_inp}
\title{Function to build SHAW .inp file.}
\usage{
shaw_inp(
  model_dir,
  iversion = "Shaw3.0",
  mtstep = 0,
  iflagsi = 1,
  inph20 = 0,
  mwatrxt = 0,
  outputs = c("profile", "soil temperature", "soil water", "soil liquid",
    "energy balance", "water balance", "frost"),
  output_freq = 24,
  site_file = paste0(basename(model_dir), ".sit"),
  inp_file = paste0(basename(model_dir), ".inp"),
  moi_file = paste0(basename(model_dir), ".moi"),
  temp_file = paste0(basename(model_dir), ".tem"),
  gro_file = paste0(basename(model_dir), ".gro"),
  wea_file = paste0(basename(model_dir), ".wea"),
  sink_file = paste0(basename(model_dir), ".sink")
)
}
\arguments{
\item{model_dir}{This is the directory where all inputs will be written.}

\item{iversion}{Version of SHAW.Valid versions include Shaw2.3 through Shaw3.0.}

\item{mtstep}{Time step: 0 = hourly, 1 = daily, 2 = same intervals as NRPDT. Defaults to 0.}

\item{outputs}{Vector with names of desired outputs: options are "profile", "soil temperature", "soil water",
"soil liquid", "soil matric", "canopy temperature", "canopy humidity", "snow temperature", "energy balance", "water balance",
"water flow", "root extract", "lateral", "frost", "salt", "soil solution"}

\item{output_freq}{Frequency in hours for outputs specified in outputs. Same value applied to all.}

\item{inp_file}{Name of .inp file. Defaults to basename of the model directory.}

\item{time_step}{Number of hours per time step; must be evenly divisible into 24.}
}
\description{
This function allows you to build SHAW files based on desired parameters.
}
\examples{
shaw_inputs(model_dir = "/Volumes/research_storage2/arctic/arctic_point_modeling/SHAW/test_shaw",
start = "2000-01-01", start_hour = 0, end = "2000-12-31")

}
\keyword{SHAW.}
\keyword{hydrology,}
