% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shaw_modify_sit.R
\name{shaw_modify_sit}
\alias{shaw_modify_sit}
\title{Function to build SHAW input files based on parameters}
\usage{
shaw_modify_sit(
  site,
  start = NULL,
  start_hour = NULL,
  end = NULL,
  lat_deg = NULL,
  lat_min = NULL,
  slope = NULL,
  aspect = NULL,
  elev = NULL,
  soils_df = NULL,
  nres = NULL,
  roughness = NULL,
  measurement_height = NULL,
  ponding = NULL,
  itype = NULL,
  pintrcp = NULL,
  xangle = NULL,
  canalb = NULL,
  tcrit = NULL,
  min_stomatal_resistance = NULL,
  stomatal_exp = NULL,
  critical_leaf_water = NULL,
  leaf_resist = NULL,
  root_resist = NULL,
  plant_height = NULL,
  dchar = NULL,
  clumping = NULL,
  biomass = NULL,
  LAI = NULL,
  rooting = NULL,
  residue_thickness = NULL,
  residue_weight = NULL,
  residue_fraction = NULL,
  residue_albedo = NULL,
  lower_bc_temp = NULL,
  nsalt = NULL,
  iwrc = NULL,
  ivlcbc = NULL,
  itmpbc = NULL,
  dry_soil_albedo = NULL,
  soil_albedo_exponent = NULL,
  residue_coefficient = NULL,
  residue_krb = NULL,
  isnotemp = NULL,
  snotemp = NULL,
  snow_roughness = NULL,
  Kst = NULL,
  Tlower = NULL,
  Tupper = NULL,
  Topt = NULL,
  Kvpd = NULL,
  r = NULL,
  mcanflg = NULL,
  istomate = NULL,
  canma = NULL,
  canmb = NULL,
  wcandt = NULL,
  hrnoon = NULL,
  nplants = NULL,
  nsp = NULL,
  nsolutes = NULL,
  error_tol = NULL,
  time_step = NULL,
  nrchang = NULL,
  gmcdt = NULL,
  debugging_seq = NULL,
  gro_file = NULL,
  run_name = NULL
)
}
\arguments{
\item{start}{This is the start date, formatted as a y-m-d character.}

\item{start_hour}{An integer (0-24) indicating start hour.}

\item{end}{This is the end date, formatted as a y-m-d. Note there is no end hour.}

\item{soils_df}{Possible variable names are: "depth", "sand", "silt", "clay", "rock", "organic", "bulk_density", "ksat", "klat", "air_entry",
"theta_sat", "n", "theta_res", "l", "alpha", "brooks_corey_lambda", "asalt", "dispersion_coefficient", "b".}

\item{nres}{Number of desired residue nodes if the residue does not change over the simulation (required). if 0, no residue. Set to 1 if residue properties change.}

\item{itype}{This controls plant type. Everything in Lines F-1 to F1-NPLANT should be a vector of length nplants.}

\item{mcanflg}{Flag controlling options for input of plant growth curves and node spacing.  (0 = no plant growth, i.e. leaf area index and plant height are constant for simulation, and model will determine node spacing within the canopy; 1 = input files for plant growth are specified for each plant and model will determine node spacing within the canopy; 2 = no plant growth and allows user to input spacing and parameters of  plant canopy layers; 3 = input files for plant growth are specified for each plant and the user can specify desired heights above ground surface for canopy nodes.)  Option 3 is intended for subsequent comparison with measurements of temperature and humidity at specified heights within the canopy.
If MCANFLG = 1, plant height, dchar, clumping, biomass, LAI, and rooting are not used, but are taken from the .gro file.}

\item{istomate}{Flag to select option for computing stomatal resistance; 1 = default computation of stomatal resistance as a function of leaf water potential; 2 = option for additional controls on stomatal conductance using Stewart-Jarvis type functions for solar radiation, air temperature and vapor pressure deficit.}

\item{gro_file}{Defaults to sharing a name with the site. However, if nplants > 1 and MCANFLG == 1 this can be a vector of gro file names for each plant.}

\item{model_dir}{This is the directory where all inputs will be written.}

\item{existing_site_file}{The existing site file}

\item{new_site_file}{The new site file.}
}
\description{
This function allows you to build SHAW files based on desired parameters.
Most parameters default to NULL, and nothing in the site file changes. If you change any of the parameters from NULL, the parameter will change in the new site file.
}
\keyword{SHAW.}
\keyword{hydrology,}
