% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shaw_sit.R
\name{shaw_sit}
\alias{shaw_sit}
\title{Function to build SHAW input files based on parameters}
\usage{
shaw_sit(
  model_dir,
  start,
  start_hour,
  end,
  lat_deg,
  lat_min,
  slope,
  aspect,
  elev,
  soils_df,
  nres,
  roughness,
  measurement_height,
  ponding,
  itype,
  pintrcp,
  xangle,
  canalb,
  tcrit,
  min_stomatal_resistance,
  stomatal_exp,
  critical_leaf_water,
  leaf_resist,
  root_resist,
  plant_height,
  dchar,
  clumping,
  biomass,
  LAI,
  rooting,
  residue_thickness,
  residue_weight,
  residue_fraction,
  residue_albedo,
  lower_bc_temp,
  nsalt = 0,
  iwrc = 3,
  ivlcbc = 1,
  itmpbc = 1,
  dry_soil_albedo = 0.25,
  soil_albedo_exponent = 1,
  residue_coefficient = 50000,
  residue_krb = 6,
  isnotemp = 1,
  snotemp = 0,
  snow_roughness = 0.15,
  Kst = 100,
  Tlower = -999,
  Tupper = 999,
  Topt = 20,
  Kvpd = 0.5,
  r = 0.5,
  mcanflg = 0,
  istomate = 2,
  canma = -53.72,
  canmb = 1.32,
  wcandt = 0,
  hrnoon = 12,
  nplants = 1,
  nsp = 0,
  nsolutes = 0,
  error_tol = 0.01,
  time_step = 1,
  nrchang = 0,
  gmcdt = 0,
  debugging_seq = c(0, 0, 0, 0, 0, 0),
  site_file = paste0(basename(model_dir), ".sit"),
  gro_file = paste0(basename(model_dir), ".gro"),
  run_name = basename(model_dir)
)
}
\arguments{
\item{model_dir}{This is the directory where all inputs will be written.}

\item{start}{This is the start date, formatted as a y-m-d character.}

\item{start_hour}{An integer (0-24) indicating start hour.}

\item{end}{This is the end date, formatted as a y-m-d. Note there is no end hour.}

\item{soils_df}{Possible variable names are: "depth", "sand", "silt", "clay", "rock", "organic", "bulk_density", "ksat", "klat", "air_entry",
"theta_sat", "n", "theta_res", "l", "alpha", "brooks_corey_lambda", "asalt", "dispersion_coefficient", "b".}

\item{nres}{Number of desired residue nodes if the residue does not change over the simulation (required). if 0, no residue. Set to 1 if residue properties change.}

\item{itype}{This controls plant type. Everything in Lines F-1 to F1-NPLANT should be a vector of length nplants.}

\item{mcanflg}{Flag controlling options for input of plant growth curves and node spacing.  (0 = no plant growth, i.e. leaf area index and plant height are constant for simulation, and model will determine node spacing within the canopy; 1 = input files for plant growth are specified for each plant and model will determine node spacing within the canopy; 2 = no plant growth and allows user to input spacing and parameters of  plant canopy layers; 3 = input files for plant growth are specified for each plant and the user can specify desired heights above ground surface for canopy nodes.)  Option 3 is intended for subsequent comparison with measurements of temperature and humidity at specified heights within the canopy.
If MCANFLG = 1, plant height, dchar, clumping, biomass, LAI, and rooting are not used, but are taken from the .gro file.}

\item{istomate}{Flag to select option for computing stomatal resistance; 1 = default computation of stomatal resistance as a function of leaf water potential; 2 = option for additional controls on stomatal conductance using Stewart-Jarvis type functions for solar radiation, air temperature and vapor pressure deficit.}

\item{gro_file}{Defaults to sharing a name with the site. However, if nplants > 1 and MCANFLG == 1 this can be a vector of gro file names for each plant.}

\item{iversion}{Version of SHAW.Valid versions include Shaw2.3 through Shaw3.0.}

\item{mtstep}{Time step: 0 = hourly, 1 = daily, 2 = same intervals as NRPDT. Defaults to 0.}
}
\description{
This function allows you to build SHAW files based on desired parameters.
}
\examples{
shaw_inputs(model_dir = "/Volumes/research_storage2/arctic/arctic_point_modeling/SHAW/test_shaw",
start = "2000-01-01", start_hour = 0, end = "2000-12-31")

}
\keyword{SHAW.}
\keyword{hydrology,}
