% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shaw_temp.R
\name{shaw_temp}
\alias{shaw_temp}
\title{Function to build SHAW moisture initial temperature conditions for the soil profile.}
\usage{
shaw_temp(
  model_dir,
  soil1 = 15,
  lower_bc = NA,
  temp_obs,
  site_file = paste0(model_dir, "/", basename(model_dir), ".sit")
)
}
\arguments{
\item{model_dir}{This is the directory where all inputs will be written.}

\item{soil1}{First line of the soils.}

\item{lower_bc}{Lower boundary condition if known. Not used if NA.}

\item{temp_obs}{Observations available. Must be a data frame with columns named depth and temperature.}

\item{site_file}{This is the site file - will be used to determine saturated or residual values.}
}
\description{
This function allows you to build SHAW files based on desired parameters.
}
\examples{
observed <- data.frame(depth = c(0.1, 0.5, 1), VWC = c(0.4, 0.2, 0.1))
shaw_moi(model_dir = "/Volumes/research_storage2/arctic/arctic_point_modeling/SHAW/test_shaw",
        method = "half", iwrc = 3, soil1 = 15)

shaw_moi(model_dir = "/Volumes/research_storage2/arctic/arctic_point_modeling/SHAW/test_shaw",
        method = "observed", iwrc = 3, soil1 = 15)

}
\keyword{SHAW.}
\keyword{hydrology,}
